Move Utils to SignedRequest class.
